generate chart to fake file
