+++ Makefile
@@ -4,7 +4,7 @@ VELA_VERSION ?= 0.1.0
 GIT_COMMIT ?= git-$(shell git rev-parse --short HEAD)
 VELA_VERSION_VAR := github.com/oam-dev/kubevela/version.VelaVersion
 VELA_GITVERSION_VAR := github.com/oam-dev/kubevela/version.GitRevision
-LDFLAGS ?= "-X $(VELA_VERSION_VAR)=$(VELA_VERSION) -X $(VELA_GITVERSION_VAR)=$(GIT_COMMIT) -X main.chartTGZSource=$$(cat -) -s -w"
+LDFLAGS ?= "-X $(VELA_VERSION_VAR)=$(VELA_VERSION) -X $(VELA_GITVERSION_VAR)=$(GIT_COMMIT)"
 
 GOX      = go run github.com/mitchellh/gox
 TARGETS  := darwin/amd64 linux/amd64 windows/amd64
@@ -25,7 +25,9 @@ test: fmt vet lint
 
 # Build manager binary
 build: fmt vet lint
-	go run hack/chart/generate.go | go build -o bin/vela -ldflags ${LDFLAGS} cmd/vela/main.go
+	go run hack/chart/generate.go
+	go build -o bin/vela -ldflags ${LDFLAGS} cmd/vela/main.go
+	git checkout cmd/vela/fake/chart_source.go
 
 npm-build:
 	cd dashboard && npm run build && cd ./..
@@ -37,7 +39,8 @@ generate-source:
 	go run hack/frontend/source.go
 
 cross-build:
-	go run hack/chart/generate.go | GO111MODULE=on CGO_ENABLED=0 $(GOX) -ldflags $(LDFLAGS) -parallel=3 -output="_bin/{{.OS}}-{{.Arch}}/vela" -osarch='$(TARGETS)' ./cmd/vela/
+	go run hack/chart/generate.go
+	GO111MODULE=on CGO_ENABLED=0 $(GOX) -ldflags $(LDFLAGS) -parallel=3 -output="_bin/{{.OS}}-{{.Arch}}/vela" -osarch='$(TARGETS)' ./cmd/vela/
 
 compress:
 	( \
