+++ cmd/vela/fake/chart_source.go
@@ -1,5 +1,5 @@
 package fake
 
-// ChartSource is a base64-encoded, gzipped tarball of the charts/vela-core dir.
-// This whole file will be rewrite at build time, please don't change this file.
+// ChartSource is a base64-encoded, gzipped tarball of the default Helm chart(charts/vela-core).
+// Its value is initialized at build time. This whole file will be rewrite at that time, please don't change this file.
 var ChartSource string
