+++ hack/frontend/source.go
@@ -9,6 +9,8 @@ import (
 	"os"
 	"time"
 
+	"github.com/oam-dev/kubevela/hack/utils"
+
 	"github.com/mholt/archiver/v3"
 )
 
@@ -44,31 +46,7 @@ func PrintToFile(data string) {
 	var buffer bytes.Buffer
 	buffer.WriteString(`package fake
 var FrontendSource = "`)
-	FprintZipData(&buffer, []byte(data))
+	utils.FprintZipData(&buffer, []byte(data))
 	buffer.WriteString(`"`)
 	_ = ioutil.WriteFile("cmd/vela/fake/source.go", buffer.Bytes(), 0644)
-}
-
-// From https://github.com/rakyll/statik/blob/master/statik.go#L313
-// FprintZipData converts zip binary contents to a string literal.
-func FprintZipData(dest *bytes.Buffer, zipData []byte) {
-	for _, b := range zipData {
-		if b == '\n' {
-			dest.WriteString(`\n`)
-			continue
-		}
-		if b == '\\' {
-			dest.WriteString(`\\`)
-			continue
-		}
-		if b == '"' {
-			dest.WriteString(`\"`)
-			continue
-		}
-		if (b >= 32 && b <= 126) || b == '\t' {
-			dest.WriteByte(b)
-			continue
-		}
-		fmt.Fprintf(dest, "\\x%02x", b)
-	}
 }
