+++ hack/utils/common.go
@@ -1,0 +1,30 @@
+package utils
+
+import (
+	"bytes"
+	"fmt"
+)
+
+// From https://github.com/rakyll/statik/blob/master/statik.go#L313
+// FprintZipData converts zip binary contents to a string literal.
+func FprintZipData(dest *bytes.Buffer, zipData []byte) {
+	for _, b := range zipData {
+		if b == '\n' {
+			dest.WriteString(`\n`)
+			continue
+		}
+		if b == '\\' {
+			dest.WriteString(`\\`)
+			continue
+		}
+		if b == '"' {
+			dest.WriteString(`\"`)
+			continue
+		}
+		if (b >= 32 && b <= 126) || b == '\t' {
+			dest.WriteByte(b)
+			continue
+		}
+		fmt.Fprintf(dest, "\\x%02x", b)
+	}
+}
